+++ apps/storybook/src/stories/Advisory.stories.tsx
@@ -45,3 +45,24 @@ export const AdvisoryDefault: Story = {
     ],
   },
 };
+
+export const AdvisoryWithMisuse: Story = {
+  args: {
+    misuse: {
+      misuseCode: "schema_version_invalid",
+      misuseText: 'Schema version is invalid: "4.adlkjg.8"',
+    },
+    errors: [
+      {
+        keyword: "total",
+        keywordLocation: "header",
+        instanceLocation: "",
+        error: "total is required",
+      },
+    ],
+    warnings: [
+      { rule: "test-rule-1", details: "", description: "Warning #1" },
+      { rule: "test-rule-2", details: "", description: "Warning #2" },
+    ],
+  },
+};